Correción: - Dtos. - OfferEntity (cambio Integer a Boolean).
