+++ src/main/java/com/ccsw/bidoffice/offer/model/OfferSearchDto.java
@@ -1,86 +1,10 @@
 package com.ccsw.bidoffice.offer.model;
 
-import java.time.LocalDate;
-
 import org.springframework.data.domain.Pageable;
-
-import com.ccsw.bidoffice.opportunity_status.model.OpportunityStatusDto;
-import com.ccsw.bidoffice.opportunity_type.model.OpportunityTypeDto;
-import com.ccsw.bidoffice.sector.model.SectorDto;
 
 public class OfferSearchDto {
 
-    private Long id;
-
-    private String client;
-
-    private String name;
-
-    private SectorDto sector;
-
-    private LocalDate requestedDate;
-
-    private OpportunityTypeDto opportunityType;
-
-    private OpportunityStatusDto opportunityStatus;
-
     private Pageable pageable;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getClient() {
-        return client;
-    }
-
-    public void setClient(String client) {
-        this.client = client;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public SectorDto getSector() {
-        return sector;
-    }
-
-    public void setSector(SectorDto sector) {
-        this.sector = sector;
-    }
-
-    public LocalDate getRequestedDate() {
-        return requestedDate;
-    }
-
-    public void setRequestedDate(LocalDate requestedDate) {
-        this.requestedDate = requestedDate;
-    }
-
-    public OpportunityTypeDto getOpportunityType() {
-        return opportunityType;
-    }
-
-    public void setOpportunityType(OpportunityTypeDto opportunityType) {
-        this.opportunityType = opportunityType;
-    }
-
-    public OpportunityStatusDto getOpportunityStatus() {
-        return opportunityStatus;
-    }
-
-    public void setOpportunityStatus(OpportunityStatusDto opportunityStatus) {
-        this.opportunityStatus = opportunityStatus;
-    }
 
     public Pageable getPageable() {
         return pageable;
