+++ src/main/java/com/ccsw/bidoffice/person/model/PersonEntity.java
@@ -41,7 +41,7 @@ public class PersonEntity {
     private String grade;
 
     @Column(name = "active")
-    private Integer active;
+    private Boolean active;
 
     public Long getId() {
         return id;
@@ -115,11 +115,11 @@ public class PersonEntity {
         this.grade = grade;
     }
 
-    public Integer getActive() {
+    public Boolean getActive() {
         return active;
     }
 
-    public void setActive(Integer active) {
+    public void setActive(Boolean active) {
         this.active = active;
     }
